refactor: split client validation redaction and timeout budget
